+++ project-management-system-server/src/main/java/projectManagementSystem/repository/UserRepository.java
@@ -8,24 +8,24 @@ import org.springframework.transaction.annotation.Transactional;
 import projectManagementSystem.entity.User;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
     @Transactional
     @Modifying
     @Query("delete from User u where u.id = ?1")
-    int deleteById(int id);
+    int deleteById(long id);
 
     @Transactional
     @Modifying
     @Query("update User u set u.name = ?2 where u.id = ?1")
-    int updateUserNameById(int id, String name);
+    int updateUserNameById(long id, String name);
 
     @Transactional
     @Modifying
     @Query("update User u set u.email = ?2 where u.id = ?1")
-    int updateUserEmailById(int id, String email);
+    int updateUserEmailById(long id, String email);
 
     @Transactional
     @Modifying
     @Query("update User u set u.password = ?2 where u.id = ?1")
-    int updateUserPasswordById(int id, String password);
+    int updateUserPasswordById(long id, String password);
 }
